[receiver/kafka] Extract partition worker without behavior change
